requirements.txt
flask==2.3.3
gunicorn==21.2.0
pillow==10.0.0
qrcode==7.4.2
streamlit==1.26.0